refactoring action types
