+++ app/actions/action-type/index.js
@@ -1,10 +1,3 @@
-export const actionTypes = {
-  FAILURE: "FAILURE",
-  INCREMENT: "INCREMENT",
-  DECREMENT: "DECREMENT",
-  RESET: "RESET",
-  LOAD_DATA: "LOAD_DATA",
-  LOAD_DATA_SUCCESS: "LOAD_DATA_SUCCESS",
-  START_CLOCK: "START_CLOCK",
-  TICK_CLOCK: "TICK_CLOCK"
-};
+import * as actionTypes from "./types";
+
+export default actionTypes;
