+++ app/actions/action-type/types.js
@@ -1,0 +1,8 @@
+export const FAILURE = "FAILURE";
+export const INCREMENT = "INCREMENT";
+export const DECREMENT = "DECREMENT";
+export const RESET = "RESET";
+export const LOAD_DATA = "LOAD_DATA";
+export const LOAD_DATA_SUCCESS = "LOAD_DATA_SUCCESS";
+export const START_CLOCK = "START_CLOCK";
+export const TICK_CLOCK = "TICK_CLOCK";
