+++ app/actions/index.js
@@ -1,13 +1,4 @@
-export const actionTypes = {
-  FAILURE: "FAILURE",
-  INCREMENT: "INCREMENT",
-  DECREMENT: "DECREMENT",
-  RESET: "RESET",
-  LOAD_DATA: "LOAD_DATA",
-  LOAD_DATA_SUCCESS: "LOAD_DATA_SUCCESS",
-  START_CLOCK: "START_CLOCK",
-  TICK_CLOCK: "TICK_CLOCK"
-};
+import actionTypes from "./action-type";
 
 export function failure(error) {
   return {
