+++ app/components/ComponentWithTrans.js
@@ -5,7 +5,6 @@ export default function ComponentWithTrans() {
   return (
     <p>
       <Trans i18nKey="common:transComponent">Translate</Trans>
-      <Trans i18nKey="common:testing">Hi</Trans>
     </p>
   );
 }
