--- app/components/PureComponent.js
@@ -1,5 +1,0 @@
-import React from 'react'
-
-export default function PureComponent ({ t }) {
-  return <p>{t('common:pureComponent')}</p>
-}
