+++ app/reducer/index.js
@@ -1,4 +1,4 @@
-import { actionTypes } from "../actions";
+import actionTypes from "../actions/action-type";
 
 export const exampleInitialState = {
   count: 0,
