+++ pages/index.js
@@ -2,8 +2,6 @@ import React from "react";
 import Link from "next/link";
 import { connect } from "react-redux";
 import { compose } from "redux";
-import PureComponent from "../app/components/PureComponent";
-import ExtendedComponent from "../app/components/ExtendedComponent";
 import ComponentWithTrans from "../app/components/ComponentWithTrans";
 import { withI18next } from "../app/lib/withI18next";
 
@@ -15,8 +13,6 @@ const Test = ({ t }) => (
     <div>
       <button>{t("sample_button")}</button>
     </div>
-    <PureComponent t={t} />
-    <ExtendedComponent />
     <ComponentWithTrans />
     <Link href="/page2">
       <a>{t("link.gotoPage2")}</a>
